Ch7/SmartBot/requirements.txt
Flask
asyncio
botbuilder-core
botbuilder-ai